LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-event-target-expected.txt



FAIL CSS Overflow: event.target and event.currentTarget for ::scroll-marker is its originating element assert_equals: event.currentTarget for ::scroll-marker click is its originating element expected Element node <div id="originatingElement"></div> but got null
